refactor: Make function just union boxes
